Stop registering fallback services that are already present in Autofac
- Combined Autofac and ServiceProvider tests into ScopingContainerTestBase
